- String8Set: Adding empty. Reformat file. Make empty string splitting consistently return String8Set.Empty. Ensure this[] and Count behave correctly for String8Set.Empty.
- Remove duplicate \r trimming in BaseTabularReader (descendants should handle because they know the real line delimiters)

- Adding CsvReader and tests.
