+++ Elfie/Elfie.Test/Elfie.Test.csproj
@@ -85,8 +85,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SampleElfieStructure.V1.cs" />
     <Compile Include="SampleElfieStructure.cs" />
-    <Compile Include="Serialization\BaseTabularWriterTests.cs" />
-    <Compile Include="Serialization\TsvReaderTests.cs" />
+    <Compile Include="Serialization\BaseTabularTests.cs" />
     <Compile Include="Serialization\TsvWriterTests.cs" />
     <Compile Include="Verify.cs" />
   </ItemGroup>
